Decrease teleports and jumppads lag
This is the original teleport lag fix based on the solution from DM_Cans_SE
by Max. The fix used by Blowfish in PK++ 1.31 with IGNORE TELEPORT
CONFIRMATION still had bugs on the local server. This patch is recommended
along with the updated client. The old clients are expected to see the
previous teleport behavior with double-teloports on high pings.