Add Cfg.StopMatchOnPlayersQuit and make StopMatch vars votable
Adding new variable makes StopMatchOnTeamQuit works as before and avoid
any confusions in variable naming.
Cfg.StopMatchOnPlayersQuit is required for FFA and non-team gamemodes
because Painkiller does not have map autodownload. This fallbacks to the
standard map pool if none of the players have a custom map after a
successful voting. There is a drawback - spectators are considered
non-players so if you need to spectate the map with no players, you need
Cfg.StopMatchOnPlayersQuit = false, or both Cfg.StopMatchOnPlayersQuit
and Cfg.StopMatchOnTeamQuit = false (for team games).